+++ src/db/documents/oauth2/authorization_code.py
@@ -9,7 +9,7 @@ class OAuth2AuthorizationCodeDocument(Document):
     account_id: str = ObjectIdField(db_field="accountId")
     client_id: str = ObjectIdField(db_field="clientId")
     issued_at: datetime = DateTimeField(db_field="issuedAt")
-    expire_at: datetime = DateTimeField(db_field="expireAt")
+    expires_at: datetime = DateTimeField(db_field="expiresAt")
     redirect_uri: str = StringField(db_field="redirectUri")
     issuer: str = StringField()
     meta = {"collection": "oauth2-authorization-code"}
